Move `adder-toolbar` -> `AdderToolbar`
