+++ src/annotator/components/AdderToolbar.js
