+++ src/annotator/components/test/AdderToolbar-test.js
